refactor: dto 패키지에 있던 핵심 클래스들 도메인으로 이동
