+++ src/main/java/com/backendoori/ootw/weather/domain/ForecastResult.java
@@ -1,4 +1,4 @@
-package com.backendoori.ootw.weather.dto.forecast;
+package com.backendoori.ootw.weather.domain;
 
 import java.util.List;
 import com.backendoori.ootw.weather.util.deserializer.ForecastResultDeserializer;
