+++ src/main/java/com/backendoori/ootw/weather/util/deserializer/ForecastResultDeserializer.java
@@ -4,8 +4,8 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 import com.backendoori.ootw.weather.domain.ForecastCategory;
-import com.backendoori.ootw.weather.dto.forecast.ForecastResult;
-import com.backendoori.ootw.weather.dto.forecast.ForecastResultItem;
+import com.backendoori.ootw.weather.domain.ForecastResult;
+import com.backendoori.ootw.weather.domain.ForecastResultItem;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
